create first, easy tests using PhantomJS
- test with 1 writer, many listeners
- test with different writers writing at different times, many listenres
- test with many concurrent writers who make non-conflicting updates

Also, made the server print something when it starts so that spawned process in
node can listen to it's stdout and then initiate testing. Additionally,
introduced `pause` and `play` into the Pad javascript API. It essentially
buffers updates from the server and doesn't apply them to the state, allowing
explicit simulation of concurrent modifications.